do not use main image as cover in articles
