+++ apps/web/src/shared/infrastructure/soils-carbon-storage-service/soilsCarbonStorageApi.ts
@@ -1,17 +1,24 @@
 import {
   GetSoilsCarbonStoragePayload,
-  SoilsCarbonStorageGateway,
+  SoilsCarbonStorageGateway as ProjectSoilsCarbonStorageGateway,
   SoilsCarbonStorageResult,
-} from "@/shared/domain/gateways/SoilsCarbonStorageApi";
+} from "@/features/create-project/application/soilsCarbonStorage.actions";
+import {
+  GetSiteSoilsCarbonStoragePayload,
+  SiteSoilsCarbonStorageResult,
+  SoilsCarbonStorageGateway as SiteSoilsCarbonStorageGateway,
+} from "@/features/create-site/application/siteSoilsCarbonStorage.actions";
 import { objectToQueryParams } from "@/shared/services/object-query-parameters/objectToQueryParameters";
 
-export class SoilsCarbonStorageApi implements SoilsCarbonStorageGateway {
+export class SoilsCarbonStorageApi
+  implements SiteSoilsCarbonStorageGateway, ProjectSoilsCarbonStorageGateway
+{
   constructor() {}
 
   async getForCityCodeAndSoils({
     cityCode,
     soils,
-  }: GetSoilsCarbonStoragePayload) {
+  }: GetSoilsCarbonStoragePayload | GetSiteSoilsCarbonStoragePayload) {
     const queryString = objectToQueryParams({ cityCode, soils });
     const response = await fetch(
       `/api/carbon-storage/site-soils?${queryString}`,
@@ -19,7 +26,9 @@ export class SoilsCarbonStorageApi implements SoilsCarbonStorageGateway {
 
     if (!response.ok) throw new Error("Error while computing carbon storage");
 
-    const jsonResponse = (await response.json()) as SoilsCarbonStorageResult;
+    const jsonResponse = (await response.json()) as
+      | SoilsCarbonStorageResult
+      | SiteSoilsCarbonStorageResult;
     return jsonResponse;
   }
 }
