+++ apps/web/src/store.ts
@@ -3,12 +3,13 @@ import {
   GetSiteGateway,
   SaveProjectGateway,
 } from "./features/create-project/application/createProject.actions";
+import { SoilsCarbonStorageGateway as ProjectSoilsCarbonStorageGateway } from "./features/create-project/application/soilsCarbonStorage.actions";
 import { CreateSiteGateway } from "./features/create-site/application/createSite.actions";
+import { SoilsCarbonStorageGateway as SiteSoilsCarbonStorageGateway } from "./features/create-site/application/siteSoilsCarbonStorage.actions";
 import {
   ProjectsListGateway,
   SitesGateway,
 } from "./features/projects/application/projectsList.actions";
-import { SoilsCarbonStorageGateway } from "./shared/domain/gateways/SoilsCarbonStorageApi";
 
 import projectCreation from "@/features/create-project/application/createProject.reducer";
 import projectSoilsCarbonStorage from "@/features/create-project/application/soilsCarbonStorage.reducer";
@@ -18,7 +19,9 @@ import projectsList from "@/features/projects/application/projectsList.reducer";
 import currentUser from "@/features/users/application/user.reducer";
 
 export type AppDependencies = {
-  soilsCarbonStorageService: SoilsCarbonStorageGateway;
+  soilsCarbonStorageService:
+    | SiteSoilsCarbonStorageGateway
+    | ProjectSoilsCarbonStorageGateway;
   createSiteService: CreateSiteGateway;
   saveProjectGateway: SaveProjectGateway;
   getSiteService: GetSiteGateway;
